requirements.txt
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.1
google-generativeai==0.5.4
requests==2.32.5
pydantic==2.11.7
google-api-python-client==2.179.0

